# specs/001-backend-api-foundation/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: NALLO Backend API
  description: REST API for NALLO document management system
  version: 1.0.0
  contact:
    name: NALLO API Support
servers:
  - url: http://localhost:8000/api/v1
    description: Local development server
  - url: https://api.nallo.example.com/api/v1
    description: Production server
...
paths:
...
  /documents:
    post:
      tags:
        - Documents
      summary: Create document
      description: Create a new document
      operationId: createDocument
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateDocumentRequest'
      responses:
        '201':
          description: Document created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    get:
      tags:
        - Documents
      summary: List documents
      description: Retrieve list of documents with optional filtering
      operationId: listDocuments
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [draft, in_review, done, publish]
        - name: type
          in: query
          schema:
            type: string
            enum: [api, general, tutorial]
        - name: lang
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: List of documents
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentListResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /documents/{id}:
    get:
      tags:
        - Documents
      summary: Get document
      description: Retrieve document by ID
      operationId: getDocument
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Document retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentResponse'
        '404':
          description: Document not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    put:
      tags:
        - Documents
      summary: Update document
      description: Update existing document
      operationId: updateDocument
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateDocumentRequest'
      responses:
        '200':
          description: Document updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentResponse'
        '404':
          description: Document not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /documents/import:
    post:
      tags:
        - Documents
      summary: Import document
      description: Import document from markdown or OAS file
      operationId: importDocument
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                type:
                  type: string
                  enum: [api, general]
                version_id:
                  type: string
                  format: uuid
      responses:
        '201':
          description: Document imported successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentResponse'
        '400':
          description: Invalid file format
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /concepts/{id}/documents:
    get:
      tags:
        - Concepts
      summary: Get documents using concept
      description: Retrieve all documents that use this concept (impact analysis)
      operationId: getConceptDocuments
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: List of documents using concept
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentListResponse'
        '404':
          description: Concept not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /pages/{id}/documents:
    post:
      tags:
        - Pages
      summary: Link page to document
      description: Create DISPLAYS relationship between page and document
      operationId: linkPageToDocument
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LinkPageDocumentRequest'
      responses:
        '201':
          description: Page linked to document successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LinkResponse'
        '404':
          description: Page or document not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    CreateDocumentRequest:
      type: object
      required:
        - title
        - type
        - content
        - lang
      properties:
        title:
          type: string
        type:
          type: string
          enum: [api, general, tutorial]
        content:
          type: string
        lang:
          type: string
          pattern: '^[a-z]{2}$'
        tags:
          type: array
          items:
            type: string
    UpdateDocumentRequest:
      type: object
      properties:
        title:
          type: string
        content:
          type: string
        status:
          type: string
          enum: [draft, in_review, done, publish]
    DocumentResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        type:
          type: string
          enum: [api, general, tutorial]
        status:
          type: string
          enum: [draft, in_review, done, publish]
        title:
          type: string
        lang:
          type: string
        content:
          type: string
        summary:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    DocumentListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/DocumentResponse'
        total:
          type: integer
        limit:
          type: integer
        offset:
          type: integer
...
    LinkPageDocumentRequest:
      type: object
      required:
        - document_id
      properties:
        document_id:
          type: string
          format: uuid
    LinkResponse:
      type: object
      properties:
        relation_id:
          type: string
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: object